Finalize rotation and facing code.
This was painful. Now it's over. I am free.

Still some work to be done - the "top facing" texture is a placeholder that I used to debug all rotations. Optimally we want to have the actual mark of the machine to rotate based on the facing and rotation (same for Positional Machines that have their mark on the top)